Fix potential for redirect loops
Some actions, such as projects#show, result in a redirect to another page.
Unfortunately, if the user didn't have permission to visit the target page,
then they could be redirected to the redirecting page ad infinitum.

Instead, we now boot all unauthorized accesses back to the root page. In
reality, users shouldn't encounter 401 errors anyway, so the impact is minimal.