Implement WebDAV support for projects, tasks and attachments.
This change exports the projects that a user owns or is assigned to under
the path prefix /files/. Using an appropriate WebDAV client (ex. the one
built into Mac OS X, Windows, or GNOME), you can create, edit, or delete
attachments, get a LaTeX-formatted version of the content for each task,
and edit the markdown body directly.